Fix open_price and close_price datatype issue

+++ db/migrate/03_create_investments.rb
@@ -3,7 +3,7 @@ class CreateInvestments < ActiveRecord::Migration[5.1]
   def change
     create_table :investments do |t|
       t.string :name
-      t.decimal :account_balance
+      t.integer :account_balance
     end
   end
